lamnt98 commit move all manage class to classService 01/07
