+++ project/src/main/java/com/example/webDemo3/controller/ClassController.java
@@ -17,19 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ClassController {
 
     @Autowired
-    private ClassListService classListService;
-
-    @Autowired
-    private GiftedClassListService giftedClassListService;
-
-    @Autowired
-    private AddClassService addClassService;
-
-    @Autowired
-    private AddGiftedClassService addGiftedClassService;
-
-    @Autowired
-    private EditClassService editClassService;
+    private ClassService classService;
 
     /**
      * kimpt142
@@ -41,7 +29,7 @@ public class ClassController {
     public ClassTableResponseDto getClassTale()
     {
         ClassTableResponseDto responseDto = new ClassTableResponseDto();
-        responseDto = classListService.getClassTable();
+        responseDto = classService.getClassTable();
         return  responseDto;
     }
 
@@ -55,7 +43,7 @@ public class ClassController {
     public GiftedClassResponseDto getGiftedClassList()
     {
         GiftedClassResponseDto responseDto = new GiftedClassResponseDto();
-        responseDto = giftedClassListService.getGiftedClassList();
+        responseDto = classService.getGiftedClassList();
         return  responseDto;
     }
 
@@ -68,7 +56,7 @@ public class ClassController {
     @PostMapping("/addclass")
     public AddClassResponseDto addNewClass(@RequestBody AddClassRequestDto model)
     {
-        return addClassService.addNewClass(model);
+        return classService.addNewClass(model);
     }
 
     /**
@@ -80,7 +68,7 @@ public class ClassController {
     @PostMapping("/addgifftedclass")
     public MessageDTO addNewGiftedClass(@RequestBody AddGiftedClassRequestDto model)
     {
-        return addGiftedClassService.addGiftedClass(model);
+        return classService.addGiftedClass(model);
     }
 
     /**
@@ -92,7 +80,7 @@ public class ClassController {
     @PostMapping("/editclass")
     public MessageDTO editClass(@RequestBody EditClassRequestDto model)
     {
-        return editClassService.editClass(model);
+        return classService.editClass(model);
     }
 
     /**
@@ -104,6 +92,6 @@ public class ClassController {
     @PostMapping("/viewclassinfor")
     public ClassInforResponseDto viewClassInfor(@RequestBody ClassInforRequestDto model)
     {
-        return classListService.getClassInfor(model);
+        return classService.getClassInfor(model);
     }
 }
