+++ project/src/main/java/com/example/webDemo3/service/ClassService.java
@@ -3,6 +3,7 @@ package com.example.webDemo3.service;
 import com.example.webDemo3.dto.*;
 import com.example.webDemo3.dto.request.AddClassRequestDto;
 import com.example.webDemo3.dto.request.AddGiftedClassRequestDto;
+import com.example.webDemo3.dto.request.ClassInforRequestDto;
 import com.example.webDemo3.dto.request.EditClassRequestDto;
 
 public interface ClassService {
@@ -16,4 +17,6 @@ public interface ClassService {
     MessageDTO addGiftedClass(AddGiftedClassRequestDto model);
 
     MessageDTO editClass(EditClassRequestDto model);
+
+    ClassInforResponseDto getClassInfor(ClassInforRequestDto model);
 }
