--- project/src/main/java/com/example/webDemo3/service/GiftedClassListService.java
@@ -1,10 +1,0 @@
-package com.example.webDemo3.service;
-
-import com.example.webDemo3.dto.GiftedClassResponseDto;
-
-/**
- * kimpt142 - 29/6
- */
-public interface GiftedClassListService {
-    GiftedClassResponseDto getGiftedClassList();
-}
